+++ jkube-kit/enricher/generic/src/test/java/org/eclipse/jkube/enricher/generic/ProjectLabelEnricherTest.java
@@ -26,6 +26,9 @@ import org.eclipse.jkube.kit.enricher.api.JKubeEnricherContext;
 import io.fabric8.kubernetes.api.model.KubernetesList;
 import io.fabric8.kubernetes.api.model.KubernetesListBuilder;
 import io.fabric8.kubernetes.api.model.LabelSelector;
+import io.fabric8.kubernetes.api.model.apps.DaemonSet;
+import io.fabric8.kubernetes.api.model.apps.DaemonSetBuilder;
+import io.fabric8.kubernetes.api.model.apps.DaemonSetSpec;
 import io.fabric8.kubernetes.api.model.apps.Deployment;
 import io.fabric8.kubernetes.api.model.apps.DeploymentBuilder;
 import io.fabric8.kubernetes.api.model.apps.DeploymentSpec;
@@ -182,7 +185,8 @@ class ProjectLabelEnricherTest {
   }
 
   @Test
-  void create_withNoConfiguredVersion_shouldAddDefaultVersionInSelector() {
+  @DisplayName("StatefulSet selector.matchLabels must not include version (immutable field, fails on version bump)")
+  void create_shouldNotAddVersionToStatefulSetSelector() {
     // Given
     KubernetesListBuilder builder = new KubernetesListBuilder().withItems(new StatefulSetBuilder().build());
 
@@ -196,17 +200,48 @@ class ProjectLabelEnricherTest {
         .extracting(StatefulSetSpec::getSelector)
         .extracting(LabelSelector::getMatchLabels)
         .asInstanceOf(InstanceOfAssertFactories.MAP)
-        .containsEntry("version", "version");
+        .containsEntry("group", "groupId")
+        .containsEntry("app", "artifactId")
+        .containsEntry("provider", "jkube")
+        .doesNotContainKey("version");
   }
 
   @Test
-  void create_withConfiguredVersion_shouldAddConfiguredVersionInSelector() {
+  @DisplayName("DaemonSet selector.matchLabels must not include version (immutable field, fails on version bump)")
+  void create_shouldNotAddVersionToDaemonSetSelector() {
     // Given
-    properties.setProperty("jkube.enricher.jkube-project-label.version", "0.0.1");
-    KubernetesListBuilder builder = new KubernetesListBuilder().withItems(new StatefulSetBuilder().build());
+    KubernetesListBuilder builder = new KubernetesListBuilder().withItems(new DaemonSetBuilder().build());
 
     // When
     projectLabelEnricher.create(PlatformMode.kubernetes, builder);
+
+    // Then
+    DaemonSet daemonSet = (DaemonSet) builder.buildFirstItem();
+    assertThat(daemonSet)
+        .extracting(DaemonSet::getSpec)
+        .extracting(DaemonSetSpec::getSelector)
+        .extracting(LabelSelector::getMatchLabels)
+        .asInstanceOf(InstanceOfAssertFactories.MAP)
+        .containsEntry("group", "groupId")
+        .containsEntry("app", "artifactId")
+        .containsEntry("provider", "jkube")
+        .doesNotContainKey("version");
+  }
+
+  @Test
+  @DisplayName("StatefulSet: version stays on the (mutable) template labels but not on the (immutable) selector")
+  void createAndEnrich_statefulSet_keepsVersionOnTemplateButNotOnSelector() {
+    // Given
+    KubernetesListBuilder builder = new KubernetesListBuilder().withItems(new StatefulSetBuilder()
+        .withNewMetadata().endMetadata()
+        .withNewSpec()
+        .withNewTemplate().withNewMetadata().endMetadata().endTemplate()
+        .endSpec()
+        .build());
+
+    // When
+    projectLabelEnricher.create(PlatformMode.kubernetes, builder);
+    projectLabelEnricher.enrich(PlatformMode.kubernetes, builder);
 
     // Then
     StatefulSet statefulSet = (StatefulSet) builder.buildFirstItem();
@@ -215,7 +250,11 @@ class ProjectLabelEnricherTest {
         .extracting(StatefulSetSpec::getSelector)
         .extracting(LabelSelector::getMatchLabels)
         .asInstanceOf(InstanceOfAssertFactories.MAP)
-        .containsEntry("version", "0.0.1");
+        .containsEntry("app", "artifactId")
+        .doesNotContainKey("version");
+    assertThat(statefulSet.getSpec().getTemplate().getMetadata().getLabels())
+        .containsEntry("app", "artifactId")
+        .containsEntry("version", "version");
   }
 
   @Test